- This commit is not ready yet as it still missing implemnetation for the pow and log but it contains implementation the exp but not tested yet

+++ benchmark/pow.cpp
@@ -32,14 +32,14 @@ BENCHMARK(BM_SMArrayPow_2D);
 // Large 2D array benchmark with variable size
 static void BM_SMArrayPow_Large(benchmark::State &state) {
     const int N = state.range(0);
-    auto arr = sm::empty<float>(N, N);
+    auto arr = sm::empty<int>(N, N);
     // Fill the array with sample values
     for (int i = 0; i < N; ++i)
         for (int j = 0; j < N; ++j)
             arr(i, j) = i + j + 1;
 
     for (auto _: state) {
-        constexpr float exponent = 2;
+        constexpr int exponent = 2;
         auto result = sm::pow(arr, exponent);
         benchmark::DoNotOptimize(result);
         benchmark::ClobberMemory();
